fix: CompressionTrees diverge from the actual array children  (#1430)
Fixed the following misalignments:
- Sparse compressor would try to compress both indices and values
`like.child(0)`
- FoR compressor didn't always return an FoR array
- Chunked compressor had entirely the wrong children in its
CompressionTree
- FSST compressor would compress the varbin offsets but didn't reflect
it in the tree
- Dict compressor switched codes and values (because we do codes first,
values second *everywhere* except in the DictArray children themselves,
where they are indexes 1 and 0 respectively). Switched the DictArray
children, which is a backcompat break, but figured we're better doing
the "natural" ordering that people seemingly expect / we do everywhere
else in the code base

Also improved the CompressionTree display impl to use TreeDisplay, which
is much nicer & more useful for debugging. And I removed the
`panic_in_result_fn` lint, since we ban `panic`, but assertions are in
fact quite nice sometimes.