feat: implement SizeLimitedStream for backpressure (#1477)
Implement a simple size-limited queue. Similar to `FuturesUnordered`,
except it caps the total amount bytes that outstanding tasks will
materialize into memory.

The goal is to use this in #1466 to give us a backpressure mechanism
that relies on memory rather than an arbitrary # of requests.

There are explanatory doc comments on the `SizeLimitedStream` and
`SizedFut` types.