Count into_canonical invocations (#1615)
This is proving pretty useful for debugging.

I don't want to make any speedy decisions like "we should cache
canonicalized arrays inside ArrayData", although we could, because at
the moment this is highlighting places we're not doing things correctly.

For example, FilterMask::clone was supposed to cache the indices/slices
iterator to speed up filtering... but it only populates these in the
child mask. Meaning if the parent is cloned again, it will be
re-computed.

To see the output, I've been running our usual benchmarks with `-F
canonical_counter` to enable this logging.