fix: BitPackedArray enforces can only be built over non-negative values (#1705)
Following up from #1699.

In the previous PR we allowed signed arrays to be bit-packed directly.
However, we did not explicitly reject arrays with negative values. We
**need** to do this because it is critical for ensuring we have fast
`search_sorted` over BitPacked data with patches, only when the patches
sort to the right-most side of the array can we do efficient binary
search.

I've added explicit preconditions that values are non-negative, and made
the BitPackedArray constructor unsafe to make it clear to callers that
they must explicitly check this themselves (the recommended safe way to
create a BPA is via the `BPA::encode()` method, which returns an Error
if there are negative values).