fix: list_to_arrow uses proper PType (#1776)
Part of #1749, specifically related to
https://github.com/spiraldb/vortex/issues/1749#issuecomment-2566109410

ListArray can have offsets of any integral type. We already had the code
to handle casting them to either i32 or i64, but we were not using the
already-cast offsets when constructing the result.

I've verified locally that the sample Vortex file from the GH issue can
load to Arrow.