Segment Alignment (#1883)
Instead of storing these in the segment map, we could store these in the
Array flatbuffer, and then can pass them into `SegmentReader::get(id,
alignment)`.

The problem with this is that it would require two round trips to load a
FlatLayout. One to fetch the array flatbuffer, then we know the
alignment of the data buffers, and then a second RT to fetch the data
buffers. Unless of course we store data buffer alignment in the
FlatLayout metadata? That could work actually...

FLUP: move ArrayData flatbuffer into the IPC definition as an
ArrayMessage. Move row_count out of ArrayParts.

Open Questions:
* Who should control / configure compression + encryption? I don't think
this should be handled inside the segment writer, since it doesn't know
what type of segment it is (data, flatbuffer, etc.). Instead, if the
layouts themselves request encryption / compression, then we can
configure different properties for different layouts. e.g. different
encryption key for each column.