Rescope expression then the identity is a struct dtype (#1887)
Defines a transform on expression with an scope type of struct(..). The
transform splits the expr in to a series of expressions that can be
evaluated only using a single field of the struct, and a combination
expr which combines the series of smaller expr into the original expr.

This is used https://github.com/spiraldb/vortex/pull/1893.

---------

Co-authored-by: Nicholas Gates <nick@nickgates.com>