Vortex Layouts - Drivers (#1914)
Introduces the idea of a `driver` as a pluggable way of orchestrating
I/O and CPU based work.

An `ExecDriver` takes a `Stream<Future<T>>` and returns a `Stream<T>`.
The abstraction means we can implement drivers that use the current
runtime thread (Inline), spawn the work onto an explicit Runtime, or
spawn the work onto a thread pool.

The `IoDriver` takes a stream of segment requests and has to resolve
them. For now, we have a FileIoDriver that assumes the bytes are laid
out on disk (and therefore coalescing is desirable), so it pulls
requests off the queue, coalesces them, then launches concurrent I/O
requests to resolve them.

The VortexFile remains generic over the I/O driver, meaning the
send-ness of the resulting ArrayStream is inferred based on the
send-ness of the I/O driver (in the case of FileDriver, that is the
send-ness of the VortexReadAt).

Part of #1676