Add ValidateVTable (#1979)
This provides encodings with the chance to reject buffers or metadata
during construction, either from owned data or during deserialization.

I ran into a few issues during alignment changes where we panic later on
trying to coerce the buffer to the correct alignment.

We could... in theory.. pass an owned value here and allow the arrays to
align their own buffers?