Array Metadata (#1985)
The current ArrayMetadata system was put together somewhat quickly and
uses `serde` + Flexbuffers to serialize all metadata.

The end state looks like this:

- [ ] An array has a fixed 8 bytes of metadata. If it needs more, it
should use a buffer (made possible by #1743)
- [ ] Rkyv can optionally be used to help with serde for these bytes.
- [ ] No eager deserialization of metadata is performed, although arrays
should validate metadata in the `ValidateVTable` (see #1979).

To support 8 byte metadata, we need to:
- [ ] Move scalars and scalar values out of metadata (e.g.
ConstantArray)
- [ ] Move shift from FoR into BitPacking (this is a bit cheeky, it's
not strictly necessary, but FoR is then left with a 8-byte PValue for
metadata, and shifting feels like it should live in BitPacking anyway?)
- [ ] All other metadata should easily fit into 8 bytes.