Emulate blob storage latency in benchmarks (#2049)
The nvme devices we use for benchmarks (both in the cloud and locally)
are incredibly fast and have incredibly low latency, which makes them
forgiving in exactly the way real object storage isnt.
This is an attempt at allowing us to emulate the performance of real
cloud system in a more reproducible way, the constants and distributions
are not set in stone and we can defiantly change them around, especially
for ranged read where we have more specific numbers from AnyBlob.