Fix scalar eq, cmp, hash, etc. (#2091)
Previously we were being a little bit lazy and delegating to
ScalarValue. We've been bitten by this in the past!

It is never safe to delegate to ScalarValue since it doesn't have enough
information to know what "equals" means, i.e. it's missing the
surrounding scalar's DType.

This PR removes PartialEq, PartialOrd, and Hash from ScalarValue, in
favor of forcing us to wrap the value into a Scalar and then compare
each dtype-specific scalar type.