feat: use decompress_into to reduce memory needed for FSST canonical (#2184)
We added the Decompressor::decompress_into unsafe method recently which
allows us to pass a decode buffer to the decompressor.

We use the already existing uncompressed_lens child to calculate the
size of the decode buffer. In most cases this will likely lead to a ~4x
reduction in memory usage for decoded FSST, as we were allocating 8x but
the actual compression is usually ~2x.

---------

Co-authored-by: Robert Kruszewski <github@robertk.io>