feat: lift comparison to constant null to the top-level (#2233)
This seems to be in the spirit of handling common cases at the top-level
and making assumptions in compute methods. This might avoid
decompression into arrow for arrays that do not currently support
comparisons. That said, I doubt this appears in the wild. I imagine all
query engines can do this optimization themselves.