fix: PR compress benchmarks (#2419)
PR compression benchmarks fail because the PR results do not have a
"storage" key. They indeed do not use any hard drive or object storage.
Moreover, those benchmarks don't use our `QueryMeasurement` /
`GenericMeasurement` structs, so they don't get a `null` storage key.

This ensures that if _either_ the PR results or the base results are
missing a storage key, we treat it as NA.

If we try to compare results from before the "storage" key to after the
"storage" key, this will not error but the right join will only include
the PR's results. The table will look very similar to the one I'm
removing here.