feat: btrblocks compressor v0 (#2247)
This PR is meant to replicate the compressor logic of btrblocks.
BtrBlocks compressor has separate schemes for each DType. Currently
starting with the integer codecs since that has the most complex search
tree with the current sampling compressor.

### Overview

- We have a new `Scheme` trait, which is the interface for implementing
compressors. Schemes are specialized based on the input array +
statistics. For example, `IntegerScheme` is a Scheme that operates over
`PrimitiveArray`s that have been summarized with an `IntegerStats`, and
`FloatScheme` is one that specializes over PrimitiveArray with
`FloatStats`.
- Every Scheme can be provided a stats instance and use it to estimate
the compression ratio that recursively compressing would produce. For
example, for integers there is a `DictScheme` which based on the
distinct values and run count, will try and estimate what Dict -> RLE +
BP codes would produce.
- Schemes can also be applied directly via their `compress` method,
which will eagerly perform recursive compression

On our PBI compression benchmarks, `Compress::choose_scheme` does seem
to only take ~1% of walltime, which matches the BtrBlocks paper result.

The goal is to FLUP and wire this compressor in as the block-level
compressor we apply when writing chunked layouts.