feat: Refining`Statistics`, `StatsSet` and related functionality (#2522)
This PR attempts to refine the `Statistics` trait and behavior. It
removes that trait, moves much of its functionality to the `StatsSetRef`
type, and tries to reduce the amount of codepaths and locations you have
to read when working on/with array statistics.

Things I like:
1. less weird `impl dyn X + '_`.
1. Hopefully less codepaths to do things, but I'm honestly not sure yet
😴
1. Better separation of what's core "statistics" behavior and whats the
extension (`StatsProvider` and `StatsProviderExt`).
1. I think some APIs are more consistent now with regards to what they
accept.
1. Removes a `std::sync::RwLock` from our API.
1. `parking_lot` supposedly faster and just panics when the lock is
poisoned.


Things I don't like
1. ~Need to import all kinds of traits, easily fixed if we ignore that
clippy lint.~
1. ~Relationship between `StatsSet` and `StatsSetRef` is weird.~ I think
this is better now.
1. `parking_lot` comes with all kinds of issues around wasm that we need
to figure out, but many of our dependencies already ship it so it should
be solvable.
1. ~I don't like traits with one implementation, it this has a bunch of
those.~
1. The new `ArrayStats` is weird and weirdly named.
1. Ideally, the `compute_` functions and `set` also would have a trait
and extensions, but I couldn't figure out a good one, so I'll leave that
for a future PR.

Things left to do:
- [x] Remove the old traits and functionality to make sure I didn't miss
anything.
- [x] tests/perf
- [x] Move/rename things to make it actually grokable.
- [ ] Read parking_lot/wasm stuff again, just to make sure I understand
it and see what needs to happen.