Context -> Registry (#2529)
Instead of handing out globally unique u16 encoding IDs, we use string
EncodingIDs for both arrays and layouts.

The footer of a vortex file lists the encoding IDs that are referenced
and these are resolved from a Registry into a Context during read. The
Context stores a Vec of encodings that can be indexed by the u16 that is
local to the file.