fix: SparseArray#search_sorted handles non null fill value that is between patch values (#2585)
Need to add some comments to the code since it took me a while to figure
out this flow. The code before was naive and didn't acocunt for fill
value that falls in between patch values which could still constitute a
sorted sparse array, i.e. `[A, fill, fill, fill, B]`