Lint all toml files using Taplo (#2455)
[taplo](https://taplo.tamasfe.dev/) is a pretty nice tool for doing all
kinds of things with TOML, including the vscode extension and a TOML
LSP. It also has a formatter so we can stop having changes that move
random things around in `Cargo.toml`.

I also added `cargo-sort` because the sorting functionality is nicer
moving the dependencies around and nothing else (as you can see from
this PR, taplo is quiet aggressive).

---------

Co-authored-by: Robert Kruszewski <github@robertk.io>