Dynamic compute fns (#2639)
Add the ability to define and dispatch compute at runtime.

Each compute function is now an `ArcRef<dyn ComputeFn>`, these are no
longer hard-coded and can be defined by users.

The entry-point of a compute function can ask an array for a "kernel"
for the specific compute function. This allows arrays to provide
implementation kernels for any compute function, even those it doesn't
have a compile-time dependency on.

It also allows arrays to define dynamic kernels runtime, for example,
DictArray can define a single kernel to push-down compute for all
element-wise operators.