fix: use new Compressor::rebuild_from for FSST pushdown (#2708)
The previous implementation of FSST comparison pushdown relied on
rebuilding a compressor by reinserting symbols one-by-one into the
CompressorBuilder, and then building it.

That doesn't work, for reasons described in the description at
https://github.com/spiraldb/fsst/pull/84.

We use the new `rebuild_from` API on fsst compressor to build a new
compressor that is guaranteed to preserve symbol table ordering, and
thus guarantee equal compression outputs.