Invert compression ratio in `BtrBlocksCompressedWriter` (#2839)
compression ratio is often
[defined](https://en.wikipedia.org/wiki/Data_compression_ratio#Definition)
as `before / after` ("higher is better"), that's also what we do in the
BtrBlocks compressor, so it seems like making it consistent across the
codebase makes sense. ~I thought there's another place we use the term,
but I couldn't find it.~ I think it was when we store the previous
chunk's compression ratio, got it before merging.