Increase default TPC-H benchmark iterations (#2840)
Seems like TPC-H is quite noisy, even on disk. Doubling the iterations
will hopefully help a bit with the noise while not making a huge
difference in overall benchmark runtime (When running on disk - it might
even still be slower than building the binary).
Also changed a bunch of string-based `default_value` to
`default_value_t` which takes on the post-parsing value.