Store only VortexMetrics until we emit them from DF (#2885)
This is a bit of a temporary hack. Basically, it's hard for us to keep
our metrics registry in sync with the DataFusion one, so instead of
trying to do this (we were duplicating lots of metrics into an every
growing registry), we only combine them when printing them out at the
end.