Downscale indices child arrays (#2862)
Because some arrays don't go through the compressor, we still want to
make sure we narrow the codes array (which realistically should almost
always be a u16 or maybe a u32 that can be bit packed).
There are also multiple opportunities to downscale various children that
don't have to maintain their DType.

For the single-file Clickbench flavor, the size change is from 21.76GB
to 19.37 (over 10%).