fix: python publishing pipeline (#2898)
1. Stop running pytest an additional time after building the wheel. As
far as I can tell it triggers another build which is the actual code it
runs. Instead we just run `pytest` in an isolated env in CI.
2. Explicitly configure the linux compatibility we want to target.
3. Try and make better use of uv's isolation capabilities to get more
predictable outcomes with steps that don't depend on the results or
state generated by previous steps.

Some of the issues we've had are caused by a surprising behavior by
uv/maturin/pep517 that might be a bug, I've filed an issue with maturin
and hopefully I'll be able to make some progress with them.