Fix: float serde handles non 0 values getting widened (#2930)
When f32/f16 gets serialized as an integer that then gets widened it's
invalid to interpret the value as the widened type (f32 -> u32 -> u64 ->
f32) since the value doesn't contain necessary padding. If we get
unsinged integer value for a float value we assume that it has the bit
width of the bits of the float value and interpret it as such.