Prevent compression from converging into low-quality compressions (#3092)
File size before this change is about 21.18GB, after this change its
19.23GB.

The issue we found is that for large files, we tend to converge into
not-compressing, or settling for mediocre compressions. With this change
we don't keep if the encoding is canonical or otherwise non
satisfactory, we don't keep that state and try again on the next chunk.