Pull out layout flatbuffer (#3127)
For queries that exclusively use column-level stats, in theory we could
defer loading the layout flatbuffer entirely...

In any case, this makes it easier to build custom logic for deferred
flatbuffer verification over layouts (when it's not included as part of
footer validation). We could build similar logic into DType flatbuffer
if that also gets expensive.