Dynamic compute kernels (#3153)
Demonstrate that we can dynamically dispatch compute kernels based on
properties of the compute function.

This is a little bit weird, because it does seem to involve making some
assumptions about the inputs to the function. I wonder if we should
codify these as conventions perhaps? It also suggests that maybe the
CastFn should pass a DType in options rather than inputs so we can
assume that all inputs are "data"