fix: Adjust the constant optimization in `list_contains` (#3211)
The constant optimization in `list_contains` was incorrect (and could
cause a crash like #3210), because it:
1. produced a `ConstantArray` with the length of the `elements` array,
rather than with the length of the input `ListArray`.
2. produced a `ConstantArray(true)` when all `elements` matched, without
correlating with the `offsets` (which might contain empty lists).

This change adjusts the constant optimization to only trigger if no
elements match, and to produce a `ConstantArray` of the appropriate
length.

Fixes #3210.