Remove unused take_into function (#3236)
Believe it or not, this is never actually called anywhere by Vortex
scans.

Happy to look into adding `invoke_into` to all compute functions in the
future if it's warranted, but for now it breaks the compute function API
for something that isn't used anywhere.

The benchmarks degrade, but they exercise a code path that isn't used in
practice. So not too bothered about them at the moment.