Add simple decimal compression (#3244)
Add a `DecimalBytePartsEncoding`.

This array contains between 1..=4 children, where the first contains the
right most none zero i64 bytes and other children contain the rest of
the bytes as u64.

For an `i128` decimal it can be split into parts:
`|| 127..64 | 63..0 || ==> {127..64, 63..0}`

The msp can be a nullable and holds the validity of the whole encoding,
where as the other children are non-nullable u64s.

Currently only |lower_parts| == 0.

We can follow up with a different encoding which splits the decimal by
integer and fractional part.

---------

Signed-off-by: Joe Isaacs <joe.isaacs@live.co.uk>