VTable Trait (#3273)
Arrays to encodings are like instances to classes in OOP. This PR fixes
a couple of things to make this more true:

* Encodings can now hold heap-allocated data, in the same way that
arrays can. This makes it possible to e.g. tie a Python class object to
an encoding by holding a reference to the Python object.
* Both encoding and array (class and instance) functions now appear on
the _same_ vtable trait, with encoding and array functions being passed
their respective heap-allocated data as the first argument. This trait
is implemented for a "<Encoding>VTable" struct, and so is hidden from
users of the "<Encoding>Encoding" and "<Encoding>Array" structs.
* Because the vtable trait is now both independent _and_
non-object-safe, we can auto-implement lots of things (see e.g.
ValidityVTableForValidityChild which automatically delegates all
validity calls to a child array).
* ArrayImpl and its ugly underscore-prefixed functions are gone because
behaviours are not implemented directly on the array structs (which
previously exposed this behaviour as semi-public API). We therefore
retain the ability to perform pre- and post-condition checks when we
`impl Array for ArrayAdapter<VTable>`.
* `dyn Array` and `dyn Encoding` are now sealed traits, meaning we
control that all implementations must go via a vtable trait and
therefore we can enforce the pre- and post-conditions, vs previously
where someone could work around them. worked

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>