duckdb: switch to https://github.com/cameron314/concurrentqueue (#3279)
This PR switches to the lock-free queue used in the DuckDB Vortex
extension to https://github.com/cameron314/concurrentqueue. Besides
removing an additional vcpkg managed dependency (DuckDB repo includes
this 3rd party lock-free queue), we need to use a queue which allocates
on push instead of throwing an error in case the capacity is exhausted.

Signed-off-by: Alexander Droste <alexander.droste@protonmail.com>