Abstract access to ArrayChildren during decode (#3287)
Instead of exposing ArrayParts (which requires a flatbuffer array
representation), this PR changes the VTable to accept a `&dyn
ArrayChildren` that provides an abstract accessor to decode children.

This will allow us to use the same build function to remove
replace_children from the vtable, that way sharing the same validation
logic as the decode.

---------

Signed-off-by: Nicholas Gates <nick@nickgates.com>